fix make http request inconsistencies

+++ integration_tests/debug.sh
@@ -1,17 +1,49 @@
 #!/bin/bash
 
+if [ $1 = "-h" ]; then
+ echo "Help for ./debug.sh"
+ printf "\n1. To run all integration tests of 'x' protocol:\n"
+ printf " \$ ./debug.sh http\n\n"
+ printf "2. To run all integration tests of 'x' protocol that contains 'y' in template name:\n"
+ printf " \$ ./debug.sh http self\n\n"
+ printf "3. To run all integration tests of 'x' protocol that contains 'y' in template name and pass extra args to nuclei:\n"
+ printf " \$ ./debug.sh http self -svd -debug-req\n\n"
+ printf "nuclei binary is created everytime script is run but integration-test binary is not"
+ exit 0
+fi
+
+
 echo "::group::Build nuclei"
 rm nuclei 2>/dev/null
 cd ../v2/cmd/nuclei
 go build .
 mv nuclei ../../../integration_tests/nuclei 
-echo "::endgroup::"
+echo -e "::endgroup::\n"
 cd ../../../integration_tests
-pwd
-./integration-test -protocol $1 -template $2
+
+cmdstring=""
+
+if [ -n "$1" ]; then
+ cmdstring+=" -protocol $1 "
+fi
+
+if [ -n "$2" ]; then
+ cmdstring+=" -template $2 "
+fi
+
+# Parse any extra args that are directly passed to nuclei
+if [ -n $debugArgs ]; then
+ export DebugExtraArgs="${@:3}"
+fi
+
+echo "$DebugExtraArgs"
+
+echo -e "::group::Run Integration Test\n"
+./integration-test $cmdstring
 
 if [ $? -eq 0 ]
 then
+  echo -e "::endgroup::\n"
   exit 0
 else
   exit 1
